move NegociarAsignaturasBehaviour to legacy
